requirements.txt



python-magic
python-docx
flask
requests


